repair: fix progress percent counting and display
* calculate the actual progress not mean of the lower level progresses
* add percent_failed field in addition to percent_success in the server response
* increase resolution from int to float

Sample output

```
╭─────────────────┬──────────────────────────────╮
│ Status          │ ERROR                        │
│ Cause           │ shard 1 error; shard 0 error │
│ Start time      │ 06 Feb 19 15:16:15 CET       │
│ End time        │ 06 Feb 19 15:17:47 CET       │
│ Duration        │ 1m32s                        │
│ Progress        │ 70.56% / 1.31%               │
│ Datacenters     │ [dc1 dc2]                    │
├─────────────────┼──────────────────────────────┤
│ test_repair     │ 70.56% / 1.31%               │
│ test_repair_dc2 │ 0.00%                        │
╰─────────────────┴──────────────────────────────╯
╭─────────────────┬───────┬────────────────┬───────────────┬─────────────────┬───────────────╮
│ test_repair     │ shard │ progress       │ segment_count │ segment_success │ segment_error │
├─────────────────┼───────┼────────────────┼───────────────┼─────────────────┼───────────────┤
│ 192.168.100.11  │ 0     │ 71.01% / 1.31% │ 1521          │ 1080            │ 20            │
│ 192.168.100.11  │ 1     │ 70.12% / 1.31% │ 1526          │ 1070            │ 20            │
│ 192.168.100.12  │ 0     │ 0.00%          │ 0             │ 0               │ 0             │
│ 192.168.100.13  │ 0     │ 0.00%          │ 0             │ 0               │ 0             │
├─────────────────┼───────┼────────────────┼───────────────┼─────────────────┼───────────────┤
│ test_repair_dc2 │ shard │ progress       │ segment_count │ segment_success │ segment_error │
╰─────────────────┴───────┴────────────────┴───────────────┴─────────────────┴───────────────╯
```

This introduces a temporary regression that makes the total percent
decrease when switching hosts. This is understood since segment count
may increase. This will fix itself when #817 is done.

Fixes #837

+++ repair/metrics.go
@@ -124,7 +124,7 @@ func (u *progressMetricsUpdater) Update(ctx context.Context) {
 				"keyspace": unit.Unit.Keyspace,
 				"host":     n.Host,
 				"shard":    "",
-			}).Set(float64(n.PercentComplete))
+			}).Set(n.PercentComplete)
 		}
 	}
 
@@ -136,7 +136,7 @@ func (u *progressMetricsUpdater) Update(ctx context.Context) {
 			"keyspace": unit.Unit.Keyspace,
 			"host":     "",
 			"shard":    "",
-		}).Set(float64(unit.PercentComplete))
+		}).Set(unit.PercentComplete)
 	}
 
 	// aggregated total progress
@@ -146,7 +146,7 @@ func (u *progressMetricsUpdater) Update(ctx context.Context) {
 		"keyspace": "",
 		"host":     "",
 		"shard":    "",
-	}).Set(float64(p.PercentComplete))
+	}).Set(p.PercentComplete)
 }
 
 func (u *progressMetricsUpdater) Stop() {
